backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
motor
passlib
python-jose
pydantic[email]
bcrypt
